+++ vite.config.ts
@@ -1,8 +1,12 @@
 import { defineConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
+import { quasar, transformAssetUrls } from "@quasar/vite-plugin";
 
 export default defineConfig({
   resolve: { alias: { "@": "/src" } },
-  plugins: [vue()],
+  plugins: [
+    vue({ template: { transformAssetUrls } }),
+    quasar({ sassVariables: "src/quasar-variables.sass" }),
+  ],
   base: "./", //process.env.NODE_ENV === "production" ? "/Distance-Incremental-Rewritten/" : "/",
 });